+++ vertx-mongo-client/src/test/java/io/vertx/ext/mongo/DistinctTest.java
@@ -40,7 +40,7 @@ public class DistinctTest extends MongoTestBase {
     JsonObject config = getConfig();
     mongoClient = MongoClient.createNonShared(vertx, config);
     CountDownLatch latch = new CountDownLatch(1);
-    dropCollections(latch);
+    dropCollections(mongoClient, latch);
     awaitLatch(latch);
   }
 
@@ -53,7 +53,7 @@ public class DistinctTest extends MongoTestBase {
   @Test
   public void testDistinctInteger() {
     String collection = randomCollection();
-    insertDocs(collection, 10, onSuccess(inserted -> {
+    insertDocs(mongoClient, collection, 10, onSuccess(inserted -> {
       mongoClient.distinct(collection, "num", Integer.class.getName(), onSuccess(distincted -> {
         assertEquals(1, distincted.size());
         assertEquals(new Integer(123), distincted.getInteger(0));
@@ -66,7 +66,7 @@ public class DistinctTest extends MongoTestBase {
   @Test
   public void testDistinctString() {
     String collection = randomCollection();
-    insertDocs(collection, 10, onSuccess(inserted -> {
+    insertDocs(mongoClient, collection, 10, onSuccess(inserted -> {
       mongoClient.distinct(collection, "foo", String.class.getName(), onSuccess(distincted -> {
         assertEquals(10, distincted.size());
         testComplete();
@@ -78,8 +78,19 @@ public class DistinctTest extends MongoTestBase {
   @Test
   public void testDistinctBadResultClass() {
     String collection = randomCollection();
-    insertDocs(collection, 10, onSuccess(inserted -> {
+    insertDocs(mongoClient, collection, 10, onSuccess(inserted -> {
       mongoClient.distinct(collection, "foo", Object.class.getName(), onFailure(failure -> {
+        testComplete();
+      }));
+    }));
+    await();
+  }
+
+  @Test
+  public void testDistinctBatchBadResultClass() {
+    String collection = randomCollection();
+    insertDocs(mongoClient, collection, 10, onSuccess(inserted -> {
+      mongoClient.distinctBatch(collection, "foo", Object.class.getName(), onFailure(failure -> {
         testComplete();
       }));
     }));
@@ -89,7 +100,7 @@ public class DistinctTest extends MongoTestBase {
   @Test
   public void testDistinctStringBadFormat() {
     String collection = randomCollection();
-    insertDocs(collection, 10, onSuccess(inserted -> {
+    insertDocs(mongoClient, collection, 10, onSuccess(inserted -> {
       mongoClient.distinct(collection, "foo", Integer.class.getName(), onFailure(failure -> {
         testComplete();
       }));
@@ -100,7 +111,7 @@ public class DistinctTest extends MongoTestBase {
   @Test
   public void testDistinctUnexistentString() {
     String collection = randomCollection();
-    insertDocs(collection, 10, onSuccess(inserted -> {
+    insertDocs(mongoClient, collection, 10, onSuccess(inserted -> {
       mongoClient.distinct(collection, "unexist", String.class.getName(), onSuccess(distincted -> {
         assertEquals(0, distincted.size());
         testComplete();
@@ -114,7 +125,7 @@ public class DistinctTest extends MongoTestBase {
     String collection = randomCollection();
     int numDocs = 10;
     CountDownLatch latch = new CountDownLatch(numDocs);
-    insertDocs(collection, numDocs, onSuccess(inserted -> {
+    insertDocs(mongoClient, collection, numDocs, onSuccess(inserted -> {
       mongoClient.distinctBatch(collection, "foo", String.class.getName(), onSuccess(distincted -> {
         assertNotNull(distincted);
         latch.countDown();
@@ -123,66 +134,4 @@ public class DistinctTest extends MongoTestBase {
     awaitLatch(latch);
   }
 
-  protected void dropCollections(CountDownLatch latch) {
-    // Drop all the collections in the db
-    mongoClient.getCollections(onSuccess(list -> {
-      AtomicInteger collCount = new AtomicInteger();
-      List<String> toDrop = getOurCollections(list);
-      int count = toDrop.size();
-      if (!toDrop.isEmpty()) {
-        for (String collection : toDrop) {
-          mongoClient.dropCollection(collection, onSuccess(v -> {
-            if (collCount.incrementAndGet() == count) {
-              latch.countDown();
-            }
-          }));
-        }
-      } else {
-        latch.countDown();
-      }
-    }));
-  }
-
-  protected List<String> getOurCollections(List<String> colls) {
-    List<String> ours = new ArrayList<>();
-    for (String coll : colls) {
-      if (coll.startsWith("ext-mongo")) {
-        ours.add(coll);
-      }
-    }
-    return ours;
-  }
-
-  protected String randomCollection() {
-    return "ext-mongo" + TestUtils.randomAlphaString(20);
-  }
-
-  protected void insertDocs(String collection, int num, Handler<AsyncResult<Void>> resultHandler) {
-    if (num != 0) {
-      AtomicInteger cnt = new AtomicInteger();
-      for (int i = 0; i < num; i++) {
-        JsonObject doc = createDoc(i);
-        mongoClient.insert(collection, doc, ar -> {
-          if (ar.succeeded()) {
-            if (cnt.incrementAndGet() == num) {
-              resultHandler.handle(Future.succeededFuture());
-            }
-          } else {
-            resultHandler.handle(Future.failedFuture(ar.cause()));
-          }
-        });
-      }
-    } else {
-      resultHandler.handle(Future.succeededFuture());
-    }
-  }
-
-  protected JsonObject createDoc(int num) {
-    return new JsonObject().put("foo", "bar" + (num != -1 ? num : "")).put("num", 123).put("big", true).putNull("nullentry").
-            put("arr", new JsonArray().add("x").add(true).add(12).add(1.23).addNull().add(new JsonObject().put("wib", "wob"))).
-            put("date", new JsonObject().put("$date", "2015-05-30T22:50:02Z")).
-            put("object_id", new JsonObject().put("$oid", new ObjectId().toHexString())).
-            put("other", new JsonObject().put("quux", "flib").put("myarr",
-                    new JsonArray().add("blah").add(true).add(312)));
-  }
 }
