+++ vertx-mongo-client/src/test/java/io/vertx/ext/mongo/MongoClientTest.java
@@ -15,7 +15,7 @@ public class MongoClientTest extends MongoClientTestBase {
     JsonObject config = getConfig();
     mongoClient = MongoClient.createNonShared(vertx, config);
     CountDownLatch latch = new CountDownLatch(1);
-    dropCollections(latch);
+    dropCollections(mongoClient, latch);
     awaitLatch(latch);
   }
