+++ vertx-mongo-client/src/test/java/io/vertx/ext/mongo/MongoClientWithObjectIdTest.java
@@ -18,7 +18,7 @@ public class MongoClientWithObjectIdTest extends MongoClientTestBase {
     config.put("useObjectId", true);
     mongoClient = MongoClient.createNonShared(vertx, config);
     CountDownLatch latch = new CountDownLatch(1);
-    dropCollections(latch);
+    dropCollections(mongoClient, latch);
     awaitLatch(latch);
   }
